# frontend/src/imports/dbguard-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DBGuardAI",
    "description": "Collector snapshot intake, approved knowledge retrieval, and human-reviewed hardening proposals.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/snapshots/{snapshot_id}/assessment": {
      "get": {
        "summary": "Get Assessment Report",
        "description": "Evaluate a snapshot against the control registry and return findings.\n\nThis endpoint performs deterministic assessment of a snapshot using the\noffline ASSESS evaluation engine. It returns pass/fail/gap status for\neach control.",
        "operationId": "get_assessment_report_api_v1_snapshots__snapshot_id__assessment_get",
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/snapshots": {
      "post": {
        "summary": "Upload Snapshot",
        "description": "Validate and store an immutable legacy bundle or native Phase 1 snapshot.",
        "operationId": "upload_snapshot_api_v1_snapshots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CollectorBundleV020"
                  },
                  {
                    "$ref": "#/components/schemas/SnapshotV030"
                  }
                ],
                "title": "Bundle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/snapshots/{snapshot_id}": {
      "get": {
        "summary": "Get Snapshot Context",
        "description": "Return the safe, normalized context used by the proposal agent.",
        "operationId": "get_snapshot_context_api_v1_snapshots__snapshot_id__get",
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/proposals/validate-and-render": {
      "post": {
        "summary": "Validate And Render Proposal",
        "description": "Validate a HERMES proposal and render only reviewed SQL templates.\n\nThis is the trusted boundary endpoint. The HERMES agent:\n1. Uses search_approved_templates to find eligible templates\n2. Uses search_approved_knowledge to find evidence\n3. Selects a template + parameters + reasoning\n4. Submits here for validation and rendering\n\nThe API validates:\n- Template exists and is active\n- Parameters match template schema\n- Evidence references point to approved documents\n- Renders the template from PostgreSQL\n\nReturns rendered SQL for human DBA review.\nNo SQL is ever generated - only rendered from approved templates.",
        "operationId": "validate_and_render_proposal_api_v1_proposals_validate_and_render_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemediationProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HardenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/library/{kind}": {
      "get": {
        "summary": "Library Catalogue",
        "operationId": "library_catalogue_api_v1_library__kind__get",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Search"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/library/knowledge/{document_id}/content": {
      "get": {
        "summary": "Library Document Content",
        "operationId": "library_document_content_api_v1_library_knowledge__document_id__content_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/templates/ingest": {
      "post": {
        "summary": "Ingest Single Template",
        "description": "Ingest a single SQL hardening template with embedding.\n\nNew templates are always created as 'draft' status regardless of request status.\nApproval must happen separately through the template approval flow.",
        "operationId": "ingest_single_template_api_v1_templates_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateIngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/templates/search": {
      "get": {
        "summary": "Search",
        "description": "Search templates by semantic similarity.",
        "operationId": "search_api_v1_templates_search_get",
        "parameters": [
          {
            "name": "search_query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Search Query"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "Top K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/knowledge/documents": {
      "post": {
        "summary": "Ingest Knowledge Document",
        "description": "Ingest a draft, or an explicitly human-approved source document.",
        "operationId": "ingest_knowledge_document_api_v1_knowledge_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeIngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessmentReport": {
        "properties": {
          "snapshot_id": {
            "type": "string",
            "title": "Snapshot Id",
            "description": "Content-addressed snapshot identifier"
          },
          "evaluated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Evaluated At",
            "description": "Timestamp when evaluation completed"
          },
          "summary": {
            "$ref": "#/components/schemas/AssessmentSummary",
            "description": "Aggregate counts of finding statuses"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings",
            "description": "Detailed findings for each evaluated control"
          }
        },
        "type": "object",
        "required": [
          "snapshot_id",
          "summary"
        ],
        "title": "AssessmentReport",
        "description": "Complete assessment report for a single snapshot."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SnapshotV030": {
        "properties": {
          "envelope": {
            "additionalProperties": true,
            "type": "object",
            "title": "Envelope"
          },
          "baseline": {
            "additionalProperties": true,
            "type": "object",
            "title": "Baseline"
          },
          "checks": {
            "additionalProperties": true,
            "type": "object",
            "title": "Checks"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "envelope",
          "baseline",
          "checks"
        ],
        "title": "SnapshotV030",
        "description": "Native Phase 1 snapshot, validated against the shared contract."
      },
...
    }
  }
}